Spreadsheet: Move sheet management to a Workbook class
